requirements.txt
megengine==1.9.1
torch==1.8.1
torchvision==0.9.1
numpy==1.21.6
scikit-image==0.19.3
Pillow==9.2.0
tqdm==4.64.0
